Move public types into client directory
