+++ src/iframe.ts
@@ -10,7 +10,7 @@ import {
     RequestType,
     AddBtcAddressesRequest,
     AddBtcAddressesResult,
-} from './lib/PublicRequestTypes';
+} from '../client/PublicRequestTypes';
 import Cashlink from './lib/Cashlink';
 import { CashlinkStore } from './lib/CashlinkStore';
 import { loadBitcoinJS } from './lib/bitcoin/BitcoinJSLoader';
