+++ src/lib/paymentOptions/EtherPaymentOptions.ts
@@ -1,20 +1,13 @@
 import bigInt from 'big-integer';
-import { Currency, PaymentType, PaymentOptions } from '../PublicRequestTypes';
+import { EtherSpecifics, EtherDirectPaymentOptions } from '../../../client/PublicPaymentOptions';
+import { Currency, PaymentType } from '../../../client/PublicRequestTypes';
 import { ParsedPaymentOptions } from './ParsedPaymentOptions';
 import { toNonScientificNumberString, FormattableNumber } from '@nimiq/utils';
-
-export interface EtherSpecifics {
-    gasLimit?: number | string;
-    gasPrice?: string;
-    recipient?: string;
-}
 
 export type ParsedEtherSpecifics = Pick<EtherSpecifics, 'recipient'> & {
     gasLimit?: number;
     gasPrice?: bigInt.BigInteger;
 };
-
-export type EtherDirectPaymentOptions = PaymentOptions<Currency.ETH, PaymentType.DIRECT>;
 
 export class ParsedEtherDirectPaymentOptions extends ParsedPaymentOptions<Currency.ETH, PaymentType.DIRECT> {
     public amount: bigInt.BigInteger;
