+++ src/views/AddAccountSelection.vue
@@ -30,7 +30,7 @@ import { State } from 'vuex-class';
 import { WalletStore } from '../lib/WalletStore';
 import { DerivedAddress } from '@nimiq/keyguard-client';
 import { ParsedSimpleRequest } from '../lib/RequestTypes';
-import { Address } from '../lib/PublicRequestTypes';
+import { Address } from '../../client/PublicRequestTypes';
 import StatusScreen from '../components/StatusScreen.vue';
 import { Static } from '../lib/StaticStore';
 import { ERROR_CANCELED } from '../lib/Constants';
