+++ src/views/AddAddressLedger.vue
@@ -33,7 +33,7 @@ import StatusScreen from '../components/StatusScreen.vue';
 import IdenticonSelector from '../components/IdenticonSelector.vue';
 import { Static } from '../lib/StaticStore';
 import { ParsedSimpleRequest } from '../lib/RequestTypes';
-import { Address } from '../lib/PublicRequestTypes';
+import { Address } from '../../client/PublicRequestTypes';
 import { WalletInfo } from '../lib/WalletInfo';
 import { AccountInfo } from '../lib/AccountInfo';
 import LedgerApi, { RequestType as LedgerApiRequestType } from '@nimiq/ledger-api';
