+++ src/views/LogoutSuccess.vue
@@ -9,7 +9,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import { ParsedSimpleRequest } from '../lib/RequestTypes';
-import { SimpleResult } from '../lib/PublicRequestTypes';
+import { SimpleResult } from '../../client/PublicRequestTypes';
 import { State } from 'vuex-class';
 import { SmallPage } from '@nimiq/vue-components';
 import { WalletStore } from '@/lib/WalletStore';
