+++ src/views/SignMessage.vue
@@ -31,7 +31,7 @@ import { Component, Vue } from 'vue-property-decorator';
 import { Getter, Mutation } from 'vuex-class';
 import { SmallPage, AccountSelector, ArrowRightIcon, ArrowLeftSmallIcon } from '@nimiq/vue-components';
 import { ParsedSignMessageRequest } from '../lib/RequestTypes';
-import { RequestType, Account } from '../lib/PublicRequestTypes';
+import { RequestType, Account } from '../../client/PublicRequestTypes';
 import staticStore, { Static } from '../lib/StaticStore';
 import { WalletStore } from '@/lib/WalletStore';
 import { AccountInfo } from '@/lib/AccountInfo';
